+++ mvp-widget/src/main/java/ru/surfstudio/android/mvp/widget/view/CoreWidgetViewInterface.java
@@ -27,13 +27,9 @@ import ru.surfstudio.android.mvp.widget.scope.WidgetViewPersistentScope;
  * Интерфейс для всех кастомных вьюшек с презентером
  * <p>
  * !!!ВАЖНО!!!
- * 1) Необходимо вызвать метод init во время onCreate() Activity или onActivityCreated() Fragment
- * 2) кастомная вьюшка с презентером может быть только в статической иерархии вью,
- * то есть должна создаваться при старте экрана, и не может быть использована при
- * динамическом создании вью, в том числе внутри элементов RecyclerView
+ * <b/>При использовании в RecyclerView необходимо проставить атрибут enableManualInit == true в макете,
+ * вызвать метод init(scopeId) во время bind()
  * <p>
- * Эти ограничения связаны с большими сложностями реализации делегирования событий экрана
- * для динамических вью
  */
 
 public interface CoreWidgetViewInterface extends
@@ -51,10 +47,18 @@ public interface CoreWidgetViewInterface extends
     WidgetViewDelegate createWidgetViewDelegate();
 
     /**
-     * Необходимо вызвать метод init во время Activity.onCreate() или Fragment.onActivityCreated()
-     * или CoreWidgetView.onCreate()
+     * Устрарел, для отложенной инициализации используется init(scopeId)
      */
+    @Deprecated
     void init();
+
+    /**
+     * Для отложенной инициализации виджета с установкой скоуп id на основе данных
+     * Используется, если у виджета проставлен атрибут - enableManualInit == true
+     *
+     * @param scopeId
+     */
+    void init(String scopeId);
 
     void onCreate();
 }