+++ template/f-debug/src/main/java/ru/surfstudio/standard/f_debug/server_settings/ServerSettingsDebugActivityView.kt
@@ -4,6 +4,7 @@ import kotlinx.android.synthetic.main.activity_server_settings_debug.*
 import ru.surfstudio.android.core.mvp.activity.BaseRenderableActivityView
 import ru.surfstudio.android.template.f_debug.R
 import ru.surfstudio.standard.base_ui.provider.component.ComponentProvider
+import ru.surfstudio.standard.f_debug.injector.ui.screen.configurator.activity.ServerSettingsDebugScreenConfigurator
 import javax.inject.Inject
 
 /**
@@ -20,16 +21,20 @@ class ServerSettingsDebugActivityView : BaseRenderableActivityView<ServerSetting
 
     override fun getPresenters() = arrayOf(presenter)
 
-    override fun createConfigurator() = ComponentProvider.createActivityScreenConfigurator(intent, this::class)
+    override fun createConfigurator() = ServerSettingsDebugScreenConfigurator(intent)
 
     override fun renderInternal(sm: ServerSettingsDebugScreenModel) {
         server_settings_chuck_switch.setChecked(sm.isChuckEnabled)
+        server_settings_test_server_switch.setChecked(sm.isTestServerEnabled)
         addCheckedChangeListener()
     }
 
     private fun addCheckedChangeListener() {
         server_settings_chuck_switch.setOnCheckedChangeListener { _, _ ->
             presenter.setChuckEnabled(server_settings_chuck_switch.isChecked())
+        }
+        server_settings_test_server_switch.setOnCheckedChangeListener { _, isEnabled ->
+            presenter.setTestServerEnabled(isEnabled)
         }
     }
 }