+++ template/f-debug/src/main/java/ru/surfstudio/standard/f_debug/server_settings/reboot/RebootDebugActivityRoute.kt
@@ -2,12 +2,12 @@ package ru.surfstudio.standard.f_debug.server_settings.reboot
 
 import android.content.Context
 import android.content.Intent
-import ru.surfstudio.android.core.ui.navigation.activity.route.ActivityRoute
+import ru.surfstudio.android.core.ui.navigation.activity.route.ActivityWithResultRoute
 
 /**
  * Роут экрана перезапуска приложения
  */
-class RebootDebugActivityRoute : ActivityRoute() {
+class RebootDebugActivityRoute : ActivityWithResultRoute<Boolean>() {
     override fun prepareIntent(context: Context?): Intent {
         return Intent(context, RebootActivityDebugView::class.java)
     }